% Auto-generated: do not edit by hand
\name{''DraggableAndResizable}

\alias{''DraggableAndResizable}

\title{DraggableAndResizable component}

\description{

}

\usage{
''DraggableAndResizable(children=NULL, id=NULL, width=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

