% Auto-generated: do not edit by hand
\name{''Graph}

\alias{''Graph}

\title{Graph component}

\description{

}

\usage{
''Graph(children=NULL, id=NULL, isOpen=NULL, width=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{isOpen}{Logical. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

