% Auto-generated: do not edit by hand
\name{''Home}

\alias{''Home}

\title{Home component}

\description{

}

\usage{
''Home(children=NULL, id=NULL, isOpen=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{isOpen}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

