% Auto-generated: do not edit by hand
\name{''index}

\alias{''index}

\title{index component}

\description{

}

\usage{
''index(children=NULL, id=NULL, setGraphWidth=NULL, width=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{setGraphWidth}{}

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

